treegen

+++ src/ai/CFolder/CMainController.py
@@ -16,7 +16,7 @@ import socket
 class CMainController:
     def __init__(self, mode_p, file_src_p):
         pygame.init()
-        self.window = CWindow(width=800, height=600)
+        self.window = CWindow(width=900, height=800)
         self.my_socket = CSocket()
         self.screen = pygame.display.set_mode((self.window.width, self.window.height))
         pygame.display.set_caption("CMainController Example")
@@ -181,7 +181,10 @@ class CMainController:
                     self.view_root_obj = self.map_from_resistor_model_to_view_linear(resistor_list_p = self.resistor_manager.resistor_list)
                 elif self.model_src == "m":
                     self.model_root_obj = CFolderModel.my_instantiate_from_flat_file("CFolderModel.txt")
-                    self.view_root_obj = CMainController.map_from_model_to_view_tree(self.model_root_obj, None)
+                    self.view_root_obj = CMainController.map_from_model_to_view_tree(self.model_root_obj,None)
+                elif self.model_src == "g":
+                    #self.model_root_obj = CFolderModel.instantiate_from_filesystem("C:\\Program Files",max_depth=3)
+                    self.view_root_obj = CTreeView.instantiate_from_flat_file("TreeView.txt")
                 self.view_root_obj.calc_i_self_tree(0)
                 self.view_root_obj.print_tree(0)
                 self.align_tree_view()
@@ -201,6 +204,8 @@ class CMainController:
                         self.broadcast(msg)
                     elif self.mode == "client":
                         self.my_socket.server_conn.sendall(msg.encode())
+            self.align_tree_view()
+
         
         def add_child():    
             count = 0
@@ -376,13 +381,14 @@ class CMainController:
             self.window.draw_mouse_coordinates(pygame)
             self.window.draw_button(self.rect_x, self.rect_y, self.rect_width, self.rect_height, pygame)
             self.window.draw_line_button(self.rect_x+self.rect_width+10, self.rect_y, self.rect_width, self.rect_height, pygame)
-            if self.view_root_obj != None:                               
-                colliding_node = self.view_root_obj.find_by_single_point_tree(self.bot_player.ghost.p_x+self.bot_player.ghost.p_w/2, self.bot_player.ghost.p_y+self.bot_player.ghost.p_h/2, "ghost")
-                if colliding_node != None:
-                    print("ghost collision", colliding_node.guid)
-                    colliding_node.delete()
-                if self.bot_player != None:
-                    self.bot_player.play()                  
+            if self.view_root_obj != None:
+                if self.model_src == "g" and self.bot_player.ghost != None:
+                    colliding_node = self.view_root_obj.find_by_single_point_tree(self.bot_player.ghost.p_x+self.bot_player.ghost.p_w/2, self.bot_player.ghost.p_y+self.bot_player.ghost.p_h/2, "ghost")
+                    if colliding_node != None:
+                        print("ghost collision", colliding_node.guid)
+                        colliding_node.delete()
+                    if self.bot_player != None:
+                        self.bot_player.play()    
                 self.view_root_obj.draw_tree(self.screen, pygame, self.font)
                 if self.window.toggle_activate_lines == 1:
                     self.view_root_obj.draw_line_tree(self.screen, pygame)
@@ -407,7 +413,7 @@ if __name__ == "__main__":
     parser.add_argument("--ip", help="Server IP (client mode only)")
     parser.add_argument(
         "--model_src",
-        choices=["r", "m", "v"],
+        choices=["r", "m", "v", "g"],
         required=True,
         help="Run as socket server or client"
     )
